requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
python-jose>=3.3.0
passlib>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
